@@ -1,4 +1,5 @@
-import { Controller, Get, Param } from '@nestjs/common';
+import { Controller, Get } from '@nestjs/common';
+import { ApiOkResponse, ApiOperation } from '@nestjs/swagger';
 import { PokemonDto } from '../dto/pokemon.dto';
 import { PokemonService } from '../services';
 
@@ -7,7 +8,9 @@ export class PokemonController {
     constructor(private readonly pokemonService: PokemonService) {}
 
     @Get()
-    findAll(): PokemonDto[] {
+    @ApiOperation({ summary: 'Get a list of pokemon' })
+    @ApiOkResponse({ type: [PokemonDto] })
+    public findAll(): PokemonDto[] {
         return this.pokemonService.findAll();
     }
 }
